Move suppressions related to properties to correct class
